+++ modules/CPDoS.py
@@ -18,8 +18,8 @@ logger = configure_logger(__name__)
 
 def crawl_files(url, s, req_main, domain, custom_header, authent, human):
     try:
-        regexp1 = r'(?<=src=")(\/[^\/].+?\.(js|css))(?=")'
-        regexp2 = r'(?<=href=")(\/[^\/].+?\.(js|css))(?=")'
+        regexp1 = r'(?<=src=")(\/[^\/].+?\.(js|css|svg|png|jpg|jpeg))(?=")'
+        regexp2 = r'(?<=href=")(\/[^\/].+?\.(js|css|svg|png|jpg|jpeg))(?=")'
         #regexp3 = r'(?<=src=")(\/[^\/].+?)(?=")'
         #regexp4 = r'(?<=href=")(\/[^\/].+?)(?=")'
 
@@ -53,7 +53,7 @@ def run_cpdos_modules(url, s, req_main, domain, custom_header, authent, human):
     }
     try:
         req_main = s.get(
-            url,
+            uri,
             headers=headers,
             verify=False,
             allow_redirects=False,