+++ modules/utils.py
@@ -66,3 +66,7 @@ class Colors:
     BLUE = "\033[34m"
     CYAN = "\033[36m"
     RESET = "\033[0m"
+
+class Identify:
+    behavior = "\033[33m └── [INTERESTING BEHAVIOR]\033[0m"
+    confirmed = "\033[31m └── [VULNERABILITY CONFIRMED]\033[0m"